Fixing autoescape bug and moving sample handlers to separate directory
