+++ static/index.html
@@ -11,9 +11,9 @@
   <p>In the meantime, though, click on the links below to find out a bit
   more about the security features in action.</p>
   <ul>
-    <li><a href="/xss">Cross-Site Scripting (XSS) protection</a></li>
-    <li><a href="/xssi">Cross-Site Script Inclusion (XSSI) protection</a></li>
-    <li><a href="/xsrf">Cross-Site Request Forgery (XSRF) protection</a></li>
+    <li><a href="/examples/xss">Cross-Site Scripting (XSS) protection</a></li>
+    <li><a href="/examples/xssi">Cross-Site Script Inclusion (XSSI) protection</a></li>
+    <li><a href="/examples/xsrf">Cross-Site Request Forgery (XSRF) protection</a></li>
   </ul>
   </body>
 </html>
